backend/pyproject.toml
[project]
name = "mem-switch-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "qdrant-client>=1.9.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0",
    "pydantic>=2.6",
    "sqlalchemy>=2.0",
    "python-multipart>=0.0.9",
    "psutil>=5.9.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
